target/kvx: refactor register fields access
Summary:
This is in preparation for removal of the use of the FIELD() macro in
gen/registers.inc.h. Using this macro is redundent with the generated
REGISTERFIELDS[] array. Next version of this generated file will embed
all the register fields information in the array and will get rid of the
FIELD() declaration completely.

This commit introduces the following new macros to access register
fields information:
    KVX_FIELD_MASK
    KVX_FIELD_LENGTH
    KVX_FIELD_SHIFT
    KVX_FIELD_EX64
    KVX_FIELD_DP64

KVX_FIELD_MASK, LENGTH and SHIFT are meant to replace direct access to
R_reg_field_... macros.

KVX_FIELD_EX64 and KVX_FIELD_DP64 are used the same way as FIELD_EX64
and FIELD_DP64.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D13417