hw/kvx/boot: add support for .mppa_args info generation
Summary:
Add support to the KVX bootloader for generating the .mppa_args section
information found in some KVX ELF. This section maps to a boot structure
used by some payloads (e.g. ClusterOS).

This structure contains among others the parameters passed to the
payload. When a "-append" option is passed to QEMU, the content of the
option is parsed as a tuple of parameters and put into the structure.
Note that this parsing has some limitations compared to a classical
argument passing mechanism (e.g. with a shell) as it cannot handles
arguments which contain spaces.

This structure is generated only when the MPPA_ARGAREA_START symbol is
found into the loaded ELF.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Maniphest Tasks: T16613

Differential Revision: https://phab.kalray.eu/D13325