hw/kvx/mppa-cluster: add the "generate-mppa-argarea" option
Summary:
Add the "generate-mppa-argarea" option to the mppa-cluster machine. When
on, it instructs the bootloader to generate the .mppa-args section
structure if one is present in the loaded ELF.

This option is on by default, and can be disabled if this behavior is
not desired, or if the argument passing mechanism through the -append
option is too limiting. In this case, one can use the -device loader
option to load arbitrary blob to memory.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Maniphest Tasks: T16613

Differential Revision: https://phab.kalray.eu/D13326