hw/kvx: add structural machine description
Summary:
This commit adds the necessary infrastructure to describe a machine and
its peripherals using a static structure. For now this description can
be used to generate the DTB of the machine.

The goal is to extend this to be able to automatically instanciate QEMU
devices and connect them together.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Maniphest Tasks: T16704

Differential Revision: https://phab.kalray.eu/D13694