hw/kvx/mppa-cluster: use the structural description to generate the DTB
Summary:
This commit adds a structural description of the mppa-cluster machine
using the structures introduced with the previous. This description
allows to generate the machine DTB when none is provided by the user.

The generated DTB is functionnaly equivalent with the one from the ISS,
and allows to boot Linux and ClusterOS just fine, with the peripherals
actually implemented in the QEMU machine.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Maniphest Tasks: T16704

Differential Revision: https://phab.kalray.eu/D13695