hw/kvx/mppa-cluster: add the frequency property to the machine
Summary:
This property allows to set the core frequency and thus the one of the
DSU clock counter. We default to 1 Mhz to avoid having to handle to much
interrupts in OSes.

Note that the dsu-clock device still need to be amended to take this
value into account.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D13696