configure: fix the rpath value for LAO library
Summary:
When switching to meson/ninja, we forgot to adapt the escaping of the
rpath value for LAO library. Since make is not in use anymore, it's not
necessary to escape the '$' as we used to do.

Moreover, the leftover absolute rpath in the installed binary is very
not ideal. Tweak meson.build so that it automatically adds the absolute
path for the build version of the QEMU binary, and remove it for the
installed version, so that at the end we only have the relative path
left.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Maniphest Tasks: T16724

Differential Revision: https://phab.kalray.eu/D13807