hw/kvx/mppa-cluster: add the L2 cache controller to the DTB
Summary:
Add the L2 cache controller to the generated DTB, as disabled. Indicate
we're executing in QEMU using the "kalray,is-qemu" property on the L2
cache controller node. Linux uses this property to allow booting
without a cache controller when CONFIG_SMP is set.

Note that we can't enable the L2 cache controller just yet because the
firmware behind it (on the RM) is extensively using the NoC DMA, which
is not modelled yet.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D13932