target/kvx: optimize hardware loop translation
Summary:
The current hardware loop implementation is very costly because whenever
LE is modified, both TBs corresponding to old and new value of LE get
invalidated.

The goal of this commit is to optimize the hardware loop translation for
the common case, which is the loopdo instruction.

Rework the hardware loop translation in the following way:
  - At translate time, when encountering the last bundle before the
  current LE value, generate a call to the check_hw_loop helper.
  - At exec time, the helper start by checking that LE matches the PC
  value it received. If it matches, the hardware loop behaviour is
  simulated.

This way, we can keep the hardware loop generated code, even when LE
changes. The only case where we still need to do TB invalidation is when
LE gets modified explicitly. In that case we need to invalidate the TB
corresponding to the new LE value, in case it has been generated without
the call to the helper.

This change brings performance improvement of up to 12x on some
benchmarks.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D14189