target/kvx: replace PS.V64 with PS.MME in the MMU index
Summary:
Some payloads such as Linux turn on and off the MMU very often. The V64
bit however is untouched after being set once at boot time in basically
every payloads.

This commit puts MME in the MMU index instead of V64. This way when the
MMU is enabled/disabled, the QEMU TLB doesn't get flushed.

We still keep V64 in the TB state though, as it makes a difference when
translating some instructions.

This change brings a nice performance improvement with Linux, booting in
5.4s instead of 7s on my machine with nosmp. The general reactivity is
also improved, especially in SMP.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D14215