target/kvx: refactor kvx_update_cpu_state
Summary:
Instead of taking masks of changes for PS and SPS as parameters, take
only the previous values, prior to state change. This fixes
kvx_update_cpu_state computing the wrong previous SPS value because of a
PL change in PS. That would lead most of the time to unecessary TLB
invalidations.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D14473