target/kvx: fix a race condition on ILR register
Summary:
When an IRQ is raised on a core, the ILR register is updated and the
corresponding bit is set.

The current implementation directly updates ILR on an IRQ, which can
lead to a race with the CPU thread modifiying it with an instruction
such as WFXL. The generated code for WFXL is not atomic and can write
back an outdated version of ILR.

To avoid this issue, this commit introduces a temporary value to store
raised IRQs, instead of directly modifiying ILR. ILR is then updated in
kvx_cpu_has_work, kvx_cpu_exec_interrupt and when reading ILR, i.e. when
we're on the CPU thread and the BQL is held.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D14476