target/kvx: add support for indirect regfile load into bigint
Summary:
Add the gen_load_regfile_indirect_bigint function which performs an
indirect load from a register file into a bigint (>=128 bits).

The load is indirect, so we don't know the offset (the register index)
into the regfile at translate time. This is why the address computation
is done at run time.

Note that this function does not support loading from a register file
with side effects (such as SFRs). No helper is called to check for side
effects on the register.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Subscribers: bddinechin

Differential Revision: https://phab.kalray.eu/D15781