target/kvx: add mds_tcg_bigint_to_i64_satu bigint operation
Summary:
Add the unsigned saturation from bigint to i64 operation. This operation
takes a signed bigint as its first operand, and a saturation size (which
must be 64 bits or less). It computes the corresponding unsigned
saturated value and returns it in a TCGv_i64.

The computation is done by OR'ing together the upper parts of the bigint
(everything but the first 64 bits) and compare the result against 0.
Next, the absolute result is saturated on 64 bits, depending on the
previous result. Then it is saturated then on sat_size bits. Finally, it
is inverted if the operand was negative.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Subscribers: bddinechin

Differential Revision: https://phab.kalray.eu/D15782