hw/dma/kvx: add the errors and IRQ sub-component
This commit implements the behaviour of the errors and IRQ sub-component
of the MPPA DMA. This component tracks errors raised in the DMA (by
other sub-components), and raises the IRQ pin accordingly.

All possible errors are centralized in the KvxDmaError enum. This enum
maps directly to the DMA_IT_EN/DMA_IT_VECTOR registers. The
kvx_dma_report_error function will be called by other sub-components to
report an error within the DMA.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044