hw/dma/kvx: add the TX thread basic structure
This commit adds the preliminary implementation of the TX threads. It
introduces the TX thread state and the TX job context.

A thread has two job contexts: one is the currently running job, and
the other is the job in cache, waiting to be executed next. The cache is
filled by the the TX job queues (not implemented yet). An index points
to the running job in the job pool.

Some MMIO registers map to the thread state, while others map directly
to the current job context.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044