hw/dma/kvx: TX job queue: add error reporting
Add the error reporting function for the TX job queue. This function
will be called by the job queue logic in case of error, as well as by
the TX thread in case of errors in a job comming from the queue.

The function maps the global DMA errors to the local error bits in the
job queue error STATUS register. The error is then propagated to the DMA
error sub-component.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044