hw/dma/kvx: TX thread: add error reporting
Add the error reporting function for the TX thread. This function
will be called by the thread logic in case of error, as well as by
the TX completion queue in case of error while writing the completion
queue descriptor or the notification.

When the current job is a manual job, i.e. it is not coming from a job
queue, the function maps the global DMA errors to the local error bits
in the TX thread ERROR register. The error is then propagated to the DMA
error sub-component.

When the current job comes from a job queue, the error if forwarded to
that job queue instead of being handled locally.

In all cases, the thread is stopped so that another job can be
scheduled.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044