hw/dma/kvx: add TX completion queue logic
Add the behaviour implementation of the TX completion queue. When
a TX thread executes a "notify" instruction, it sends a signal to the TX
completion queue. In turn, the completion queue writes a completion
descriptor to memory at the address pointed by its write pointer, and
then write a notification message.

The size of the completion descriptor is defined by the FIELD_ENABLE
register. When 0, no descriptor are written to memory. When 1, only the
first two paramaters of the job are written. When 2, the full job
descriptor is written. In any case, the write pointer is incremented.

On error, the completion queue STATUS register is updated, and the queue
is stopped. The error is then forwarded to the TX thread which initiated
the completion notification.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044