hw/dma/kvx: add microcode bundles helper functions
Add kvx_dma_bundle_decode and kvx_dma_bundle_disas functions to help
with bundle decoding and disassembling.

kvx_dma_bundle_decode takes a raw bundle as parameter and builds the
corresponding DecodedBundle structure. In case of invalid bundle, false
is returned.

kvx_dma_bundle_disas can be used for debugging and tracing purposes. It
builds a string representing the decoded bundle given as parameter. The
caller must specify a separator to be placed before each instruction of
the bundle.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044