hw/dma/kvx: add the TX thread execution logic
This commit adds the TX thread execution behaviour. The main thread loop
in kvx_dma_run_tx_threads starts by checking if a job is waiting in
cache. If this is the case and the thread is idle, the cached job
becomes the running job and the thread is enabled.

Then the thread is run. It starts by checking if it has room in the
completion queue the job points to. If not, the thread is blocked and
can't run. It then enters a fetch/decode/execute loop effectively
simulating the microcode instructions (the execute function is not
implemented yet). The loop runs until completion (i.e., until the thread
get disabled), or upon reaching some run budget limit, fixed at 1024
instructions in this commit.

This limit is here to circumvent a possibly very long or buggy
microcode. In this case the execution is aborted and a new run is
scheduled in a bottom-half on the iothread.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044