hw/dma/kvx: add TX thread bundle execution logic
Add the bundle execution logic to the TX thread execution. The decoded
bundled is simulated according to the instructions it embeds. The
receive/send instructions will be added in the next commit.

The instruction simulation order matters as we don't want to write the
modifications to a temporary state. So the modifications appear in an
order such that following instructions don't read parts of the modified
state by the previous ones. This works as the DMA does not feature
cross-dependencies instructions.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16044