hw/kvx/boot: add the address space used by the bootloader in KvxBootInfo
Summary:
Until now, the default address_space_memory was used for memory accesses from the simulated bootloader.
To support multi-cluster configurations, we need to be able to specify an address space to use.
This commit allows this by adding the AS to use into the KvxBootInfo structure.

Signed-off-by: titouan HUARD <titouan.huard@kalray.eu>

Test Plan: CI

Reviewers: lmichel, O110 qemu!

Differential Revision: https://phab.kalray.eu/D16160