target/kvx: fix wrong AS used when invalidating a TB
When invalidating a translation block in tb_invalidate_bundle, the
address_space_memory address space was passed to
tb_invalidate_phys_addr. This is wrong in the general case, since the
CPU can use another AS to do its memory accesses.

Fix this by fetching the correct AS using cpu_get_address_space.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Differential Revision: D16283