gdbstub: fix get_threads_xml truncating the result
Summary:
The get_threads_xml function was using the process->target_xml buffer
to store the <threads>...</threads> XML tree, which is incorrect since
this buffer is normally used to store the target description XML.

Moreover, this buffer is 1024 bits only, limiting the number of threads
we can enumerate to about 23. Going over this value results in a
truncated XML snippet and GDB complaining about it.

Fix this by using a GString to store the XML snippet.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D16833