gdbstub: get_thread_xml: enumerate threads from all attached processes
Summary:
The get_thread_xml function was only listing the threads of the current
process. This function is called when building the response of the
Xfer:threads:read packet. Returning the threads of the current process
only confuses GDB making it thinks that other threads have disappeared.

Change this so the get_thread_xml function enumerates the threads of all
attached processes.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D16835