hw/kvx/itgen: fix the cluster offset in the MSI address
Summary:
When generating an MSI, the itgen peripheral computes the MSI address
from the corresponding IRQ configuration register. The register contains
a "cluster" field corresponding to the cluster to notify. The final
formula is as follows:

  msi_addr = 0xa00000 | ((cluster + 1) << 24) | (mailbox << 9)

We were missing the "+ 1" on the cluster. This commit adds this and fix
some comments.

Signed-off-by: Luc Michel <lmichel@kalray.eu>

Test Plan: CI

Reviewers: O110 qemu!

Differential Revision: https://phab.kalray.eu/D16862